TG-2: Restructure folders and added new files
